LoadImage added. Needs further refinement for our usage.
Unity project output was changed to windows when it was loaded. Most of the file changes are due to that.
The important additions are in LevelLoader, line 76 and forward. The new loadImage function loads files from a given filepath, and outputs a Texture2D. Getting the path to the StreamingAssets folder is mostly a matter of calling Application.dataPath. 
UseTestImgAsTexture.cs was made for testing the function.
We have a problem here, however. Unity Engine can only load JPG's and PNG's. A fix would probably be to handle the BMP with System.Drawing instead. 
Discussion needed.