Various tweaks to import export, fixed inverted y
The binary file now also contains a float scale factor value. For now, the exported value is a magic number. When imported, Polygons are created and parented to the test object (manager?), and then the parent is scaled down.
Polygon y values flipped when read: Screenspace is y positive down, but Unity space is y positive up. This resulted in mirrored polygons. This has been fixed.