Tried to fix up program, encountered serious problems. See description.
For some reason, we are hitting AccessViolationExceptions in weird places when trying to work with Bitmaps solely in memory. It is difficult to pinpoint why, but it seems to stem from memory management, as it does not occur if we work from a local file. Current suspicion lies in the PixelManager class, as it is doing memory management on images. If true, fixing this will take a lot of time... 
If we want results soon, the better course of action will be to store captured images on disk, and use the older function call that uses file paths.